requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
pillow==10.2.0
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.0
httpx==0.26.0
asyncpg==0.29.0
rapidfuzz==3.6.1
cachetools==5.3.2
